pyproject.toml
[project]
name = "valoov-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.20.0",
    "google-generativeai>=0.8.5",
    "pymongo>=4.13.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
